+++ web-wallet/src/__mocks__/mockedWalletStore.js
@@ -9,9 +9,8 @@ const content = {
   addresses,
   balance,
   currentAddress,
-  error: null,
   initialized: true,
-  isSyncing: false,
+  syncStatus: { current: 0, error: null, isInProgress: false, last: 0 },
 };
 
 const mockedWalletStore = mockReadableStore(content);