+++ web-wallet/src/lib/components/index.js
@@ -7,6 +7,7 @@ export { default as ApproveMigration } from "./ApproveMigration/ApproveMigration
 export { default as Balance } from "./Balance/Balance.svelte";
 export { default as ContractOperations } from "./ContractOperations/ContractOperations.svelte";
 export { default as ContractStatusesList } from "./ContractStatusesList/ContractStatusesList.svelte";
+export { default as CopyField } from "./CopyField/CopyField.svelte";
 export { default as DashboardNav } from "./DashboardNav/DashboardNav.svelte";
 export { default as ExecuteMigration } from "./ExecuteMigration/ExecuteMigration.svelte";
 export { default as ExistingWalletNotice } from "./ExistingWalletNotice/ExistingWalletNotice.svelte";