+++ web-wallet/src/lib/stores/stores.d.ts
@@ -27,6 +27,11 @@ type SettingsStoreContent = {
 
 type SettingsStore = Writable<SettingsStoreContent> & { reset: () => void };
 
+type GasSettings = {
+  limit: number;
+  price: number;
+};
+
 type TransactionsStoreContent = { transactions: Transaction[] };
 
 type TransactionsStore = Readable<TransactionsStoreContent>;
@@ -40,11 +45,15 @@ type WalletStoreContent = {
     maximum: number;
     value: number;
   };
+  syncStatus: {
+    isInProgress: boolean;
+    current: number;
+    last: number;
+    error: Error | null;
+  };
   currentAddress: string;
-  error: Error | null;
   initialized: boolean;
   addresses: string[];
-  isSyncing: boolean;
 };
 
 type WalletStoreServices = {
@@ -52,14 +61,19 @@ type WalletStoreServices = {
 
   clearLocalData: () => Promise<void>;
 
-  clearLocalDataAndInit: (wallet: Wallet) => Promise<void>;
+  clearLocalDataAndInit: (
+    wallet: Wallet,
+    syncFromBlock?: number
+  ) => Promise<void>;
+
+  getCurrentBlockHeight: () => Promise<number>;
 
   getStakeInfo: () => Promise<any> & ReturnType<Wallet["stakeInfo"]>;
 
   // The return type apparently is not in a promise here
   getTransactionsHistory: () => Promise<ReturnType<Wallet["history"]>>;
 
-  init: (wallet: Wallet) => Promise<void>;
+  init: (wallet: Wallet, syncFromBlock?: number) => Promise<void>;
 
   reset: () => void;
 
@@ -67,27 +81,23 @@ type WalletStoreServices = {
 
   stake: (
     amount: number,
-    gasPrice: number,
-    gasLimit: number
+    gasSettings: GasSettings
   ) => Promise<any> & ReturnType<Wallet["stake"]>;
 
-  sync: () => Promise<void>;
+  sync: (from?: number) => Promise<void>;
 
   transfer: (
     to: string,
     amount: number,
-    gasPrice: number,
-    gasLimit: number
+    gasSettings: GasSettings
   ) => Promise<any> & ReturnType<Wallet["transfer"]>;
 
   unstake: (
-    gasPrice: number,
-    gasLimit: number
+    gasSettings: GasSettings
   ) => Promise<any> & ReturnType<Wallet["unstake"]>;
 
   withdrawReward: (
-    gasPrice: number,
-    gasLimit: number
+    gasSettings: GasSettings
   ) => Promise<any> & ReturnType<Wallet["withdrawReward"]>;
 };