+++ web-wallet/src/lib/stores/walletStore.js
@@ -1,13 +1,7 @@
 import { get, writable } from "svelte/store";
-import { getKey, uniquesBy } from "lamb";
-
-/**
- * @typedef {import("@dusk-network/dusk-wallet-js").Wallet} Wallet
- */
-
-/**
- * @typedef {WalletStoreServices["getTransactionsHistory"]} GetTransactionsHistory
- */
+import { getKey, setKey, uniquesBy } from "lamb";
+import { Gas, Wallet } from "@dusk-network/dusk-wallet-js";
+import settingsStore from "./settingsStore";
 
 /** @type {AbortController} */
 let syncController;
@@ -28,9 +22,8 @@ const initialState = {
     value: 0,
   },
   currentAddress: "",
-  error: null,
   initialized: false,
-  isSyncing: false,
+  syncStatus: { current: 0, error: null, isInProgress: false, last: 0 },
 };
 
 const walletStore = writable(initialState);
@@ -55,39 +48,6 @@ const getCurrentAddress = () => get(walletStore).currentAddress;
 /** @type {(action: (...args: any[]) => Promise<any>) => Promise<void>} */
 const syncedAction = (action) => sync().then(action).finally(sync);
 
-const abortSync = () => {
-  syncPromise && syncController?.abort();
-  syncPromise = null;
-};
-
-/** @type {() => Promise<void>} */
-const clearLocalData = async () => walletInstance?.reset();
-
-/** @type {(wallet: Wallet) => Promise<void>} */
-const clearLocalDataAndInit = (wallet) =>
-  wallet.reset().then(() => init(wallet));
-
-/** @type {WalletStoreServices["getStakeInfo"]} */
-const getStakeInfo = async () =>
-  sync()
-    // @ts-expect-error
-    .then(() => walletInstance.stakeInfo(getCurrentAddress()))
-    .then(fixStakeInfo);
-
-/** @type {GetTransactionsHistory} */
-
-const getTransactionsHistory = async () =>
-  sync()
-    // @ts-expect-error
-    .then(() => walletInstance.history(getCurrentAddress()))
-    .then(uniquesById);
-
-function reset() {
-  abortSync();
-  walletInstance = null;
-  set(initialState);
-}
-
 async function updateAfterSync() {
   const store = get(walletStore);
 
@@ -97,12 +57,48 @@ async function updateAfterSync() {
   set({
     ...store,
     balance,
-    isSyncing: false,
+    syncStatus: initialState.syncStatus,
   });
 }
 
-/** @param {Wallet} wallet */
-async function init(wallet) {
+const abortSync = () => {
+  syncPromise && syncController?.abort();
+  syncPromise = null;
+};
+
+/** @type {WalletStoreServices["clearLocalData"]} */
+const clearLocalData = async () => walletInstance?.reset();
+
+/** @type {WalletStoreServices["clearLocalDataAndInit"]} */
+const clearLocalDataAndInit = (wallet, syncFromBlock) =>
+  wallet.reset().then(() => init(wallet, syncFromBlock));
+
+/** @type {WalletStoreServices["getCurrentBlockHeight"]} */
+const getCurrentBlockHeight = async () => Wallet.networkBlockHeight;
+
+/** @type {WalletStoreServices["getStakeInfo"]} */
+const getStakeInfo = async () =>
+  sync()
+    // @ts-expect-error
+    .then(() => walletInstance.stakeInfo(getCurrentAddress()))
+    .then(fixStakeInfo);
+
+/** @type {WalletStoreServices["getTransactionsHistory"]} */
+const getTransactionsHistory = async () =>
+  sync()
+    // @ts-expect-error
+    .then(() => walletInstance.history(getCurrentAddress()))
+    .then(uniquesById);
+
+/** @type {WalletStoreServices["reset"]} */
+function reset() {
+  abortSync();
+  walletInstance = null;
+  set(initialState);
+}
+
+/** @type {WalletStoreServices["init"]} */
+async function init(wallet, syncFromBlock) {
   walletInstance = wallet;
 
   const addresses = await walletInstance.getPsks();
@@ -114,7 +110,10 @@ async function init(wallet) {
     currentAddress,
     initialized: true,
   });
-  sync();
+
+  sync(syncFromBlock).then(() => {
+    settingsStore.update(setKey("userId", currentAddress));
+  });
 }
 
 /** @type {WalletStoreServices["setCurrentAddress"]} */
@@ -122,26 +121,25 @@ async function setCurrentAddress(address) {
   const store = get(walletStore);
 
   return store.addresses.includes(address)
-    ? Promise.resolve(set({ ...store, currentAddress: address })).then(sync)
+    ? Promise.resolve(set({ ...store, currentAddress: address })).then(() =>
+        sync()
+      )
     : Promise.reject(new Error("The received address is not in the list"));
 }
 
 /** @type {WalletStoreServices["stake"]} */
-
-const stake = async (amount, gasPrice, gasLimit) =>
+const stake = async (amount, gasSettings) =>
   syncedAction(() => {
     // @ts-expect-error
-    walletInstance.gasLimit = gasLimit;
-
-    // @ts-expect-error
-    walletInstance.gasPrice = gasPrice;
-
-    // @ts-expect-error
-    return walletInstance.stake(getCurrentAddress(), amount);
+    return walletInstance.stake(
+      getCurrentAddress(),
+      amount,
+      new Gas(gasSettings)
+    );
   });
 
 /** @type {WalletStoreServices["sync"]} */
-function sync() {
+function sync(from) {
   if (!walletInstance) {
     throw new Error("No wallet instance to sync");
   }
@@ -149,11 +147,37 @@ function sync() {
   if (!syncPromise) {
     const store = get(walletStore);
 
-    set({ ...store, error: null, isSyncing: true });
+    set({
+      ...store,
+      syncStatus: {
+        current: 0,
+        error: null,
+        isInProgress: true,
+        last: 0,
+      },
+    });
 
     syncController = new AbortController();
+
+    const syncOptions = {
+      from,
+      /** @type {(current: number, last: number) => void} */
+      onblock: (current, last) => {
+        set({
+          ...store,
+          syncStatus: {
+            current,
+            error: null,
+            isInProgress: true,
+            last,
+          },
+        });
+      },
+      signal: syncController.signal,
+    };
+
     syncPromise = walletInstance
-      .sync({ signal: syncController.signal })
+      .sync(syncOptions)
       .then(() => {
         if (syncController.signal.aborted) {
           throw new Error("Synchronization aborted");
@@ -161,7 +185,15 @@ function sync() {
       })
       .then(updateAfterSync)
       .catch((error) => {
-        set({ ...store, error, isSyncing: false });
+        set({
+          ...store,
+          syncStatus: {
+            current: 0,
+            error,
+            isInProgress: false,
+            last: 0,
+          },
+        });
       })
       .finally(() => {
         syncPromise = null;
@@ -172,42 +204,32 @@ function sync() {
 }
 
 /** @type {WalletStoreServices["transfer"]} */
-const transfer = async (to, amount, gasPrice, gasLimit) =>
+const transfer = async (to, amount, gasSettings) =>
   syncedAction(() => {
     // @ts-expect-error
-    walletInstance.gasLimit = gasLimit;
-
-    // @ts-expect-error
-    walletInstance.gasPrice = gasPrice;
-
-    // @ts-expect-error
-    return walletInstance.transfer(getCurrentAddress(), to, amount);
+    return walletInstance.transfer(
+      getCurrentAddress(),
+      to,
+      amount,
+      new Gas(gasSettings)
+    );
   });
 
 /** @type {WalletStoreServices["unstake"]} */
-const unstake = async (gasPrice, gasLimit) =>
+const unstake = async (gasSettings) =>
   syncedAction(() => {
     // @ts-expect-error
-    walletInstance.gasLimit = gasLimit;
-
-    // @ts-expect-error
-    walletInstance.gasPrice = gasPrice;
-
-    // @ts-expect-error
-    return walletInstance.unstake(getCurrentAddress());
+    return walletInstance.unstake(getCurrentAddress(), new Gas(gasSettings));
   });
 
 /** @type {WalletStoreServices["withdrawReward"]} */
-const withdrawReward = async (gasPrice, gasLimit) =>
+const withdrawReward = async (gasSettings) =>
   syncedAction(() => {
     // @ts-expect-error
-    walletInstance.gasLimit = gasLimit;
-
-    // @ts-expect-error
-    walletInstance.gasPrice = gasPrice;
-
-    // @ts-expect-error
-    return walletInstance.withdrawReward(getCurrentAddress());
+    return walletInstance.withdrawReward(
+      getCurrentAddress(),
+      new Gas(gasSettings)
+    );
   });
 
 /** @type {WalletStore} */
@@ -215,6 +237,7 @@ export default {
   abortSync,
   clearLocalData,
   clearLocalDataAndInit,
+  getCurrentBlockHeight,
   getStakeInfo,
   getTransactionsHistory,
   init,